DEEP-345 Remove unnecessary files and refactor code
