--- assets/src/scss/company.scss
@@ -1,1 +1,0 @@
-
