+++ src/router.dispatch.ts
@@ -1,13 +1,9 @@
 // Do Router dispatch here, i.e. map incoming routes to appropriate router
 import { Application, Request, Response } from "express";
-import indexRouter from "./routers/indexRouter";
-import userRouter from "./routers/userRouter";
-import companyRouter from "./routers/companyRouter";
+import startRouter from "./routers/startRouter";
 
 const routerDispatch = (app: Application) => {
-    app.use("/", indexRouter);
-    app.use("/user", userRouter);
-    app.use("/company", companyRouter);
+    app.use("/", startRouter);
     app.use("*", (req: Request, res: Response) => {
         res.status(404).render("partials/error_400");
     });
