--- src/routers/__utils/index.ts
@@ -1,2 +1,0 @@
-// General utility methods for the routes and handlers (if not deemed fit for the GenericHandler)
-// This file (and the __utils folder in which it resides) are likely to be redundant due to the GenericHandler.
