+++ src/routers/handlers/start/start.ts
@@ -1,8 +1,8 @@
 import { Request, Response } from "express";
-import { GenericHandler } from "./../generic";
+import { GenericHandler } from "../generic";
 import logger from "../../../lib/Logger";
 
-export class HomeHandler extends GenericHandler {
+export class StartHandler extends GenericHandler {
 
     constructor () {
         super();
@@ -14,15 +14,5 @@ export class HomeHandler extends GenericHandler {
         logger.info(`GET request for to serve home page`);
         // ...process request here and return data for the view
         return Promise.resolve(this.viewData);
-    }
-
-    // additional support method in handler
-    private supportMethod1 (): boolean {
-        return true;
-    }
-
-    // additional support method in handler
-    protected supportMethod2 (): boolean {
-        return false;
     }
 };
