--- src/views/router_views/index/home.njk
@@ -1,10 +1,0 @@
-{% extends "layouts/default.njk" %}
-
-{% block main_content %}
-
-  <div class="generic-class-name">
-    <h3>Index Router: Home page</h3>
-    <p>{{ sampleKey }}</p>
-  </div>
-
-{% endblock %}
