+++ src/views/router_views/start/start.njk
@@ -1,0 +1,9 @@
+{% extends "layouts/default.njk" %}
+
+{% block main_content %}
+
+  <div>
+    <h1 class="govuk-heading-l">Requesting an extension</h1>
+  </div>
+
+{% endblock %}
