--- src/views/router_views/user/profile.njk
@@ -1,10 +1,0 @@
-{% extends "layouts/default.njk" %}
-
-{% block main_content %}
-
-  <div class="generic-class-name">
-    <h3>User Router: User profile page</h3>
-    <p>{{ sampleKey }}</p>
-  </div>
-
-{% endblock %}
