--- src/views/router_views/user/settings.njk
@@ -1,10 +1,0 @@
-{% extends "layouts/default.njk" %}
-
-{% block main_content %}
-
-  <div class="generic-class-name">
-    <h3>User Router: User settings page</h3>
-    <p>{{ sampleKey }}</p>
-  </div>
-
-{% endblock %}
